requirements.txt
telebot
aiohttp
ollama
docstring_parser
pillow
serpapi
wyoming
scapy
